+++ src/state/lists/tokenFromList.ts
@@ -1,7 +1,7 @@
 import { Currency, Token } from '@uniswap/sdk-core'
 import { Tags, TokenInfo, TokenList } from '@uniswap/token-lists'
 
-import { isAddress } from '.'
+import { isAddress } from '../../utils'
 
 type TagDetails = Tags[keyof Tags]
 interface TagInfo extends TagDetails {
